+++ stockfish-core/CHANGELOG.md
@@ -16,6 +16,7 @@ adheres to [Semantic Versioning][semver].
 
 - `impl Default for Bitboard`
 - `impl BitOr<Color, Output = Piece> for Token`
+- `impl BitOr<Token, Output = Piece> for Color`
 
 ## [0.2.0](https://github.com/stouset/stockfish-rs/tree/stockfish-core-v0.2.0/stockfish-core) - 2023-01-16